ref(node): Move `net.ts` into `internal`
